requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
distro==1.9.0
Django==5.1.6
firebase-admin==6.6.0
gcloud==0.18.3
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.2
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.0
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.8.2
jwcrypto==1.5.6
msgpack==1.1.0
ngrok==1.4.0
oauth2client==4.1.3
openai==1.65.3
packaging==24.2
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
Pyrebase4==4.8.0
python-http-client==3.3.7
python-jwt==4.1.0
replicate==1.0.4
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
sendgrid==6.11.0
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
